requirements.txt
# requirements.txt
# Dependências para integração com Mercado Pago

# Processamento de dados
pandas>=1.5.0
openpyxl>=3.0.0

# Mercado Pago SDK
mercadopago>=2.2.0

# Geração de QR Codes
qrcode>=7.4.2
Pillow>=9.0.0

# Webhook (servidor Flask)
Flask>=2.3.0

# Variáveis de ambiente
python-dotenv>=1.0.0

# Utilidades
urllib3>=1.26.0
